src/task3/requirements.txt
python-dotenv==1.0.1
python-magic==0.4.27
